move examples into examples directory
